feat(api): update post, put, and delete routes        Updated the post route to add a variable for creating a product using the .create method based on the JSON request, as described in the comments on lines 44-51.        Updated the put request to update a product by using the ID endpoint and the .update method to select the product ID to update, allowing the use of a JSON request to update a product.        Added a variable to the delete product route using the .destroy method to enable deletion of a product by using the ID.
feat(api): update post, put, and delete routes

Updated the post route to add a variable for creating a product using the .create method based on the JSON request, as described in the comments on lines 44-51.

Updated the put request to update a product by using the ID endpoint and the .update method to select the product ID to update, allowing the use of a JSON request to update a product.

Added a variable to the delete product route using the .destroy method to enable deletion of a product by using the ID.